Fixes problem when local database pointer becomes nil during the POST _rev_diffs stage of a push replication (see TDPusher -processInbox).
The test added in previous commit on this branch reproduces this crash.

Adds ReplicationAcceptance -testSyncReplicationErrorsWhenLocalDatabaseDeleted_pullDelegateDeletes_indefinitely
which I've used to try to catch some intermittent crashes. It is
commented out so that it doesn't blow up in any CI tool
As a result of watching this test:

  * I've added a weakSelf/strongSelf mechanism to the onCompletion
block in TDPusher -processInbox. I was seeing crashes
at [strongSelf uploadBulkDocs:docsToSend changes:revsToSend] inside TDPusher -processInbox.

  * still intermittent crashes (EXC_BAD_ACCESS) in onCompletion block and am working to track them down.

Obviously, this is a temporary commit. Will be squashed or edited.