Do not declare remote URL when creating the datastore
We used to pass in a remote URL to associate the local database with
_and_ put all databases in a magic directory.  This no longer makes
sense since the local datastore can be, and is typically, replicated
against arbitrary remote stores.

This patch makes it more in line with the Core Data model where a file
systems URL is given.  That URL identify a directory where all
datastore files are kept.  It is now the responsibility of the user to
make sure that all but the last path component exists in the system.

Signed-off-by: Jimi Xenidis <jimix@pobox.com>